requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
distro==1.9.0
dnspython==2.7.0
frozenlist==1.6.0
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==3.10
jiter==0.10.0
multidict==6.4.4
openai==1.81.0
propcache==0.3.1
pyaes==1.6.1
pyasn1==0.6.1
pydantic==2.11.4
pydantic_core==2.33.2
pymongo==4.13.0
python-dotenv==1.0.0
python-telegram-bot==20.6
requests==2.32.3
rsa==4.9.1
sniffio==1.3.1
soupsieve==2.7
Telethon==1.40.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
yarl==1.20.0

